python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitflow-onchain"
version = "0.1.0"
description = "Bitcoin on-chain ingestion and metric workers for BitFlow"
requires-python = ">=3.12"
dependencies = [
  "psycopg[binary]>=3.2.0",
  "pyzmq>=26.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
